pycode/requirements.txt
# Core dependencies
langchain
langchain-core
langchain-community
langchain-experimental
langchain-huggingface
langchain-neo4j
langchain-ollama
langgraph
langchain-openai
langchain-qdrant

# Database and Graph
neo4j
neo4j-graphrag
chromadb
qdrant-client
pymongo
sqlalchemy
psycopg2-binary

# Document Processing
pymupdf
python-docx
lxml-html-clean
pillow

# # Data Processing
pandas
numpy
featuretools
spacy

# Machine Learning and Embeddings
sentence-transformers
transformers
nltk
torch

# Visualization
matplotlib
yfiles-jupyter-graphs
plotly

# Web Interface
gradio
Flask
flask-cors

# Utilities
tqdm
pydantic
typing-extensions
python-dotenv
pytest

# Search and Retrieval
tavily-python
duckduckgo-search
newspaper4k

# Jupyter
ipython
jupyter

# Scrapping pack
requests
requests-html
beautifulsoup4
firecrawl-py 
crawl4ai

# Document processing
bs4 
pypdf    
openpyxl
PyPDF2

# LLMs and Embeddings
ollama
openai
anthropic
fastembed
google-generativeai

# Agno/PhiData API
agno
fastapi
playground
uvicorn
python-multipart

# Huggingface
huggingface_hub[hf_xet]

# AI Evaluation - comet/opik
opik    
comet-ml
advertools
jinja2

# Intellliwrite
agno
qdrant-client
openai
pydantic
python-dotenv
sqlalchemy
pg8000
Flask
flask-cors
duckduckgo-search
langfuse